Create separate package for user service
