+++ internal/connection/database.go
@@ -4,6 +4,7 @@ import (
 	"os"
 	"fmt"
 	"errors"
+	
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 	slogGorm "github.com/orandin/slog-gorm"
